fix(financial): LedgerService.getAccountBalance returns int (ADR-018)
FASE F completada: RefundTest 9/9 pasando.

CAMBIO EN LedgerService::getAccountBalance:

Firma con parámetros opcionales:
  public function getAccountBalance(
      int $accountId,
      ?string $fromDate = null,
      ?string $toDate = null
  ): int  // antes float

Implementación:
  $debits = (int) $query->sum('debit_amount');
  $credits = (int) $query->sum('credit_amount');
  return (int) ($debits - $credits);

CADENA COMPLETA INT:
  DB (integer) → LedgerEntry (int cast) → LedgerService (int) → Test (int)

REGRESSION CHECK:
✅ FASE A-E: 60 tests
✅ FASE F (RefundTest): 9/9

TOTAL: 69 tests financieros pasando

Siguiente: FASE G (CrashRecoveryTest, BackupRecoveryTest, PerformanceTest)

+++ app/Modules/Accounting/Domain/Services/LedgerService.php
@@ -136,7 +136,7 @@ class LedgerService
     /**
      * Obtiene el balance de una cuenta.
      */
-    public function getAccountBalance(int $accountId, ?string $fromDate = null, ?string $toDate = null): float
+    public function getAccountBalance(int $accountId, ?string $fromDate = null, ?string $toDate = null): int  // ADR-018
     {
         $account = Account::findOrFail($accountId);
 
@@ -150,8 +150,8 @@ class LedgerService
             $query->whereHas('journalEntry', fn($q) => $q->where('entry_date', '<=', $toDate));
         }
 
-        $debits = $query->sum('debit_amount');
-        $credits = $query->sum('credit_amount');
+        $debits = (int) $query->sum('debit_amount');  // ADR-018
+        $credits = (int) $query->sum('credit_amount');  // ADR-018
 
         return $account->type->normalBalance() === 'debit'
             ? $debits - $credits
